refactored `$doesnotexist`
- removed static data
- moved methods definition to a separate module file
